refactor(state): extract pure reducers out of runtimeEventSlice + threadSlice
Mirrors the supervisor runtime/ split for the renderer slices. Pure reducer
logic (applyRuntimeEventToRuntimeState + coalesce/merge/prune/mergeContextUsage
helpers, ~500L) moves to runtimeEventReducer.ts; thread turn-timing helpers
(parseTurnIso/appendCompletedTurnIfClosed/resolveCompletedTurnAnchorItemId/
deriveTurnTiming) move to threadTurnHelpers.ts. The slice factories stay as thin
wrappers. runtimeEventSlice 933->440, threadSlice 851->739. The i18n macro user
(terminateSubAgentItem) and the runtimePersistenceDirtyListeners set stay behind
(not pure). runtimeEventSlice.test.ts (24) + appStore.test.ts (56) green.
